+++ src/settings/lib/RuleEditor/utils.js
@@ -1,0 +1,12 @@
+const addIndentToEditorRules = (rule, position) => {
+  switch (position) {
+    case 'before':
+      return ` ${rule}`;
+    case 'after':
+      return `${rule} `;
+    default:
+      return rule;
+  }
+};
+
+export default addIndentToEditorRules;